Move input & resize processing to gfx_app
